+++ internal/db/database.go
@@ -1,20 +1,31 @@
 package db
 
 import (
+	"context"
+	"fmt"
+	"log"
+	"time"
+
+	"go.mongodb.org/mongo-driver/v2/bson"
 	"go.mongodb.org/mongo-driver/v2/mongo"
 	"go.mongodb.org/mongo-driver/v2/mongo/options"
-	"log"
 )
 
 func NewMongoClient(conn string) *mongo.Client {
 	client, err := mongo.Connect(options.Client().ApplyURI(conn))
 
 	if err != nil {
-		log.Println("error here")
 		log.Fatalln(err)
 	}
 
-	log.Println("Connected to MongoDB!")
+	// Send a ping to confirm a successful connection
+	var result bson.M
+	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
+	defer cancel()
+	if err := client.Database("admin").RunCommand(ctx, bson.D{{Key: "ping", Value: 1}}).Decode(&result); err != nil {
+		panic(err)
+	}
+	fmt.Println("Pinged your deployment. You successfully connected to MongoDB!")
 
 	return client
 }
